Fixed bug in UziFirmware.executeSaveProgram()
The old version was failing to write relatively large programs. The problem
was that I was inmediately writing to the EEPROM each byte read from the
serial. And since writing to the EEPROM is slow, the serial buffer was
filling up and eventually I was losing data. Now I read from the serial as
fast as possible.

+++ c++/UziFirmware/Reader.cpp
@@ -3,7 +3,7 @@
 int32 Reader::nextLong(int16 size, bool& timeout)
 {
 	int32 result = 0;	
-	for (int16 i = size - 1; i >= 0; i--)
+	for (int32 i = size - 1; i >= 0; i--)
 	{
 		result |= ((uint32)next(timeout) << (i * 8));
 		if (timeout) return 0;
